adding find and put methods to Env
overall I found out that it would be better to check if an id has been declared by implementing a find functionaity in the hashmap itself. A bug showa up where the Env class isn't recognized for some reason. Incomplete gen functions added, and a getLexeme getter added to idNode